fattureCCSR: python rewrite of downloader and integration with minimal gui. Initial commit
